--- bot/src/main/java/com/community/tools/controller/GithubHookController.java
@@ -1,30 +1,0 @@
-package com.community.tools.controller;
-
-import com.community.tools.service.github.event.GithubEventsProcessingService;
-import lombok.AllArgsConstructor;
-import org.json.JSONObject;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@AllArgsConstructor
-@RequestMapping("/gitHook")
-public class GithubHookController {
-
-  private final GithubEventsProcessingService eventsProcessingService;
-
-  /**
-   * Method receive webhook data from GitHub.
-   *
-   * @param body event body
-   */
-  @PostMapping
-  public void getHookData(@RequestBody String body) {
-    JSONObject eventJson = new JSONObject(body);
-    eventsProcessingService.processEvent(eventJson);
-  }
-
-}
-
