+++ bot/src/main/java/com/community/tools/model/User.java
@@ -3,8 +3,10 @@ package com.community.tools.model;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import java.time.LocalDate;
 import java.util.Date;
+import java.util.Set;
 import javax.persistence.Entity;
 import javax.persistence.Id;
+import javax.persistence.OneToMany;
 import javax.persistence.Table;
 import javax.persistence.Transient;
 import lombok.Getter;
@@ -22,6 +24,8 @@ public class User {
   private String gitName;
   private LocalDate dateRegistration;
   private LocalDate dateLastActivity;
+  @OneToMany
+  private Set<Mentors> mentors;
   @JsonIgnore
   private byte[] stateMachine;
   private Integer karma = 0;
