--- bot/src/main/java/com/community/tools/service/github/event/EventHandler.java
@@ -1,11 +1,0 @@
-package com.community.tools.service.github.event;
-
-public interface EventHandler<T> {
-
-  /**
-   * Processes an object with event information of type T passed to it.
-   *
-   * @param event event info
-   */
-  void handleEvent(T event);
-}
